Added MACs
Enter button is now highlighted when enabled. Added MACs and refactored. Put AddMac() and Remove Mac() inside of Encrypt() and Decrypt(), and added Lock() and Unlock() methods. UI update to show status as an emoji.

+++ FileLockerLibrary/FileModel.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
+using System.Reflection.Metadata;
 using System.Security.Cryptography;
 using System.Text;
 using System.Threading.Tasks;
@@ -22,10 +24,23 @@ public class FileModel
     /// </summary>
     public string Password { get; set; }
 
-    /// <summary>
-    /// The encryption key used for encryption and decryption.
-    /// </summary>
-    public byte[] EncryptionKey { get; set; }
+    private byte[] encryptionKey;
+    public byte[] EncryptionKey
+    {
+        get
+        {
+            return encryptionKey;
+        }
+        set
+        {
+            if (value == null)
+                throw new ArgumentNullException("Encryption key cannot be null.");
+            if (value.Length != 32)
+                throw new ArgumentException("Invalid encryption key length. Expected 32 bytes.");
+
+            encryptionKey = value;
+        }
+    }
 
     /// <summary>
     /// The salt used in deriving the encryption key.
@@ -43,6 +58,72 @@ public class FileModel
         }
     }
 
+    public bool IntegrityStatus
+    {
+        get
+        {
+            bool output = false;
+
+            if (Mac != null && MacKey != null)
+                output = ValidateMac();
+
+            return output;
+        }
+
+        private set { IntegrityStatus = value; }
+    }
+
+    public byte[] MacKeySalt { get; set; }
+
+    private byte[] macKey;
+    public byte[] MacKey
+    {
+        get
+        {
+            return macKey;
+        }
+        set
+        {
+            if (value == null)
+                throw new ArgumentNullException("MAC key cannot be null.");
+            if (value.Length != 32)
+                throw new ArgumentException("Invalid MAC key length. Expected 32 bytes.");
+
+            macKey = value;
+        }
+    }
+
+    private byte[] mac;
+    public byte[] Mac
+    {
+        get
+        {
+            return mac;
+        }
+        set
+        {
+            mac = value;
+        }
+    }
+
+    public string DisplayName
+    {
+        get
+        {
+            string output = "";
+
+            if (EncryptionStatus == true)
+                output += "🔐";
+            else
+                output += "     ";
+
+            string fileName = FileName.Length > 50 ? FileName.Substring(0, 47) + "..." : FileName;
+            output += $" {fileName}";
+
+            return output;
+        }
+    }
+
     /// <summary>
     /// The file name without the directory path.
     /// </summary>
@@ -54,52 +135,112 @@ public class FileModel
         }
     }
 
-    /// <summary>
-    /// Encrypts the content of the file.
-    /// </summary>
-    public void Encrypt()
+    public void Lock()
+    {
+        Encrypt();
+        AddMac();
+    }
+
+    public void Unlock()
+    {
+        Decrypt();
+        RemoveMac();
+    }
+
+    private void Encrypt()
     {
         if (!File.Exists(Path))
             throw new FileNotFoundException("The file could not be found.", Path);
+        if (Password == null)
+            throw new NullReferenceException("Password must be set.");
         if (EncryptionStatus == true)
             return;
 
+        // read plaintext
         string plaintextPath = Path;
-        Path = Path + Constants.EncryptedExtension;
+        byte[] content = File.ReadAllBytes(Path);
+        Path += Constants.EncryptedExtension;
 
-        byte[] content = File.ReadAllBytes(plaintextPath);
+        // encrypt
+        EncryptionKeySalt = GlobalConfig.KeyDeriver.GenerateSalt();
+        EncryptionKey = GlobalConfig.KeyDeriver.DeriveKey(Password, EncryptionKeySalt);
         byte[] ciphertext = GlobalConfig.Encryptor.Encrypt(content, EncryptionKey);
 
+        // overwrite plaintext
         File.WriteAllBytes(Path, ciphertext);
         File.Delete(plaintextPath);
     }
 
-    /// <summary>
-    /// Decrypts the content of the file.
-    /// </summary>
-    /// <exception cref="FileNotFoundException">Thrown the file could not be found.</exception>
-    /// <exception cref="CryptographicException">Thrown when decryption fails.</exception>
-    public void Decrypt()
+    private void Decrypt()
     {
         if (!File.Exists(Path))
             throw new FileNotFoundException("The file could not be found.", Path);
+        if (Password == null)
+            throw new NullReferenceException("Password must be set.");
+        if (EncryptionKeySalt == null)
+            throw new NullReferenceException("Encryption key salt must be set.");
         if (EncryptionStatus == false)
             return;
 
         try
         {
+            // read ciphertext
             string ciphertextPath = Path;
             byte[] content = File.ReadAllBytes(ciphertextPath);
+
+            // decrypt
+            EncryptionKey = GlobalConfig.KeyDeriver.DeriveKey(Password, EncryptionKeySalt);
             byte[] plaintext = GlobalConfig.Encryptor.Decrypt(content, EncryptionKey);
 
+            // write plaintext
             Path = System.IO.Path.ChangeExtension(Path, null);
             File.WriteAllBytes(Path, plaintext);
+
+            // cleanup
+            EncryptionKeySalt = null;
             File.Delete(ciphertextPath);
         }
         catch (Exception ex)
         {
             throw new CryptographicException("Unable to decrypt.", ex);
         }
+    }
+
+    private void AddMac()
+    {
+        if (!File.Exists(Path))
+            throw new FileNotFoundException("The file could not be found.", Path);
+        if (Password == null)
+            throw new NullReferenceException("Password must be set.");
+        if (EncryptionStatus == false)
+            return;
+
+        MacKeySalt = GlobalConfig.KeyDeriver.GenerateSalt();
+        MacKey = GlobalConfig.KeyDeriver.DeriveKey(Password, MacKeySalt);
+        GlobalConfig.MacGenerator.Key = MacKey;
+
+        byte[] content = File.ReadAllBytes(Path);
+        Mac = GlobalConfig.MacGenerator.GenerateMac(content);
+    }
+
+    private bool ValidateMac()
+    {
+        if (Mac == null)
+            throw new NullReferenceException("MAC must be set.");
+        if (MacKey == null)
+            throw new NullReferenceException("MAC key must be set.");
+
+        byte[] content = File.ReadAllBytes(Path);
+        GlobalConfig.MacGenerator.Key = MacKey;
+        bool output = GlobalConfig.MacGenerator.ValidateMac(content, Mac);
+
+        return output;
+    }
+
+    private void RemoveMac()
+    {
+        Mac = null;
+        MacKeySalt = null;
     }
 
     /// <summary>
